requirements.txt
opencv_python_headless==4.1.2.30
natsort
numpy
tqdm
lmdb
scikit_image
matplotlib
torchsummary
nltk
torchvision
imageio
six
scipy
fire
Pillow
scikit_learn
horovod
editdistance
git+https://github.com/google/gin-config